// frontend/src/openapi/models/req-create-comment-body.ts
/* tslint:disable */
/* eslint-disable */
/**
 * ReadersLounge API
 * ReadersLounge API
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface ReqCreateCommentBody
 */
export interface ReqCreateCommentBody {
    /**
     * 
     * @type {number}
     * @memberof ReqCreateCommentBody
     */
    'post_id': number;
    /**
     * 
     * @type {string}
     * @memberof ReqCreateCommentBody
     */
    'content': string;
}