// frontend/src/openapi/models/post-like.ts
/* tslint:disable */
/* eslint-disable */
/**
 * ReadersLounge API
 * ReadersLounge API
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface PostLike
 */
export interface PostLike {
    /**
     * 
     * @type {number}
     * @memberof PostLike
     */
    'post_like_id': number;
}